+++ src/Redmine/Api/User.php
@@ -6,6 +6,8 @@ use Redmine\Exception;
 use Redmine\Exception\MissingParameterException;
 use Redmine\Exception\SerializerException;
 use Redmine\Exception\UnexpectedResponseException;
+use Redmine\Http\HttpFactory;
+use Redmine\Serializer\JsonSerializer;
 use Redmine\Serializer\PathSerializer;
 use Redmine\Serializer\XmlSerializer;
 
@@ -142,7 +144,7 @@ class User extends AbstractApi
      * @param int|string $id     the user id or `current` for retrieving the user whose credentials are used
      * @param array      $params extra associated data
      *
-     * @return array information about the user
+     * @return array|false|string information about the user as array or false|string on error
      */
     public function show($id, array $params = [])
     {
@@ -158,9 +160,22 @@ class User extends AbstractApi
         );
         $params['include'] = implode(',', $params['include']);
 
-        return $this->get(
-            PathSerializer::create('/users/' . urlencode(strval($id)) . '.json', $params)->getPath()
-        );
+        $this->lastResponse = $this->getHttpClient()->request(HttpFactory::makeJsonRequest(
+            'GET',
+            PathSerializer::create('/users/' . urlencode(strval($id)) . '.json', $params)->getPath(),
+        ));
+
+        $body = $this->lastResponse->getContent();
+
+        if ('' === $body) {
+            return false;
+        }
+
+        try {
+            return JsonSerializer::createFromString($body)->getNormalized();
+        } catch (SerializerException $e) {
+            return 'Error decoding body as JSON: ' . $e->getPrevious()->getMessage();
+        }
     }
 
     /**